Bugfix
Problem: The exchange.Send method was not working

Root cause: the exchange used Addr (an interface) as a key. Different concrete types which use the same interface don't match on the key

Fix: We create basicAddr and use that as a key.

ToDo: Consider making basicAddr exported and use that as the keyType instead of Addr